gpuav: Refacto instrumentation error logging
- Move specific instrumentation error loging logic to corresponding
files
- Relies on lambda, to have the flexibility to store custom error info
without having to touch the (now dead) InstrumentationErrorBlob struct
- vertex attribute fetching validation was still doing its descriptor
set logic in gpuav_instrumentation, also moved that to dedicated file
- removed function definitions advertised in gpuav_instrumentation.h
but only used in the .cpp
