+++ BUILD.gn
@@ -209,11 +209,13 @@ vvl_sources = [
   "layers/gpuav/instrumentation/gpuav_instrumentation.cpp",
   "layers/gpuav/instrumentation/gpuav_instrumentation.h",
   "layers/gpuav/instrumentation/buffer_device_address.cpp",
-  "layers/gpuav/instrumentation/buffer_device_address.h",
   "layers/gpuav/instrumentation/descriptor_checks.cpp",
   "layers/gpuav/instrumentation/descriptor_checks.h",
   "layers/gpuav/instrumentation/post_process_descriptor_indexing.cpp",
-  "layers/gpuav/instrumentation/post_process_descriptor_indexing.h",
+  "layers/gpuav/instrumentation/ray_query.cpp",
+  "layers/gpuav/instrumentation/register_validation.h",
+  "layers/gpuav/instrumentation/sanitizer.cpp",
+  "layers/gpuav/instrumentation/vertex_attribute_fetch_oob.cpp",
   "layers/gpuav/resources/gpuav_vulkan_objects.cpp",
   "layers/gpuav/resources/gpuav_vulkan_objects.h",
   "layers/gpuav/resources/gpuav_shader_resources.h",
