--- layers/gpuav/instrumentation/post_process_descriptor_indexing.h
@@ -1,23 +1,0 @@
-/* Copyright (c) 2024-2025 LunarG, Inc.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-#pragma once
-
-namespace gpuav {
-class Validator;
-class CommandBufferSubState;
-
-void RegisterPostProcessingValidation(Validator& gpuav, CommandBufferSubState& cb);
-
-}  // namespace gpuav
