+++ layers/gpuav/instrumentation/register_validation.h
@@ -19,5 +19,9 @@ class Validator;
 class CommandBufferSubState;
 
 void RegisterBufferDeviceAddressValidation(Validator& gpuav, CommandBufferSubState& cb);
+void RegisterPostProcessingValidation(Validator& gpuav, CommandBufferSubState& cb);
+void RegisterRayQueryValidation(Validator& gpuav, CommandBufferSubState& cb);
+void RegisterSanitizer(Validator& gpuav, CommandBufferSubState& cb);
+void RegisterVertexAttributeFetchOobValidation(Validator& gpuav, CommandBufferSubState& cb);
 
 }  // namespace gpuav
